Migrate from Kotlin synthetics to Jetpack view binding

+++ app/src/main/java/com/diazzerss/stocks/domain/model/Stock.kt
@@ -1,9 +1,11 @@
 package com.diazzerss.stocks.domain.model
 
+import java.math.BigDecimal
+
 data class Stock(
     val companyName: String,
     val ticker: String,
-    val price: Double,
-    val changes: Double,
+    val price: BigDecimal,
+    val changes: BigDecimal,
     val changesPercentage: String
 )
